Handling the especial case where the profile is 'default'
When an SSO profile is configured with the name 'default', the CLI creates the `[default]` entry in the configuration file, but the program is expecting to find `[profile default]`.

Although there is a known workaround for this issue, it envolves editing the configuration file and renaming the profile. A better approach is to support the default behavior of the CLI.

Since some users may have already implemented this workaround, the program was altered to support both '[default]' and [profile default]' profiles.